Initial reordering
Includes a move of the src/ data to data/, the scripts/ scripts to utils/, as well as
decoupling some things from the Galician language, and replacing the SCons build system
by a custom script.

Some of the utils (actually, all but the build util) will not work due to the changes,
and will need adjustments, but I will apply those later.

The project has now proper documentation that can be built into HTML files. So far there
are only three content pages (Quick Start Guide and reference pages about the build tool
and the language.json file format) but as new features are added, fixed or reviewed, I
will update the documentation accordingly so that it slowly grows.